BugFix: Fix to make the refactored code give approximatly the same results as previously
There was issues with the indexing and combining as suspected. There were a couple of problems. One was that the mock data had a file that looked like it was and index, it wasn't. This column has been removed from the data and the data is read with no index column specified.

The second is related. The above index was possibly causing an error that meant that in the notebook main for loop for label propagation was an index reset. This was not working correctly and was causing the combined labels which rely on the index to not be joined correctly. The index reset has been removed. If it was being used for anything else we can disscuss.